+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- fix import of components with circular dependencies
+
 ## [13.0.5-dev.26 - 2018-10-29]
 
 - fix link content generation for authored components on bit install